// src/types/theme.ts
export type ColorMap = {
  primary: string;
  secondary: string;
  surface: string;
  commonText: string;
  textPrimary: string;
  textSecondary: string;
  textMuted: string;
  hyperLink: string;
  button: string;
  buttonText: string;
  detailsBackground: string;
  errorText: string;
  errorBackground: string;
  errorSurface: string;
  success: string;
  warning: string;
  secondaryText: string;
  buttonDisable: string;
  bracketColor: string;
  border: string;
  overlay: string;
  methodBadgeBg: string;
  codeBackground: string;
  codeText: string;
  codeKey: string;
};

export type LogTypeKey = 'LOG' | 'WARN' | 'ERROR';